Specify SSDs as a count
Use the range function to generate a list of drives. This avoids the
need to manually create the list.

If the existing Terraform uses the format local-ssd-1, local-ssd-2, etc
this will have no effect on any deployed proxies.

However, if an existing Terraform uses different names for the local SSDs
such as drive-1, drive-2, does not start numbering from 1, or has gaps such
as local-ssd-1, local-ssd-3 this change will replace the proxies with a
new template.