Explicitly exclude protected directories
Instead of implicitly skipping any exports that would overwrite a protected
directory immediately terminate the start up script. This avoids starting
the NFS server with only a partial list of exports.

A new option to exclude exports from auto-discovery is now supported so
that any exports that would overwrite a protected directory can be
explicitly excluded.