Support UDP protocol
This is mainly for older clients that default to using UDP for the mount
protocol. NFS does not recommend using UDP for actual data transfer.

The reserved address for existing deployments uses the default of
GCE_ENDPOINT. To avoid breaking changes for existing deployments the
address has been set to ignore changes to the purpose.