Remove --managed-gids from RCPMOUNTDOPTS
The managed-gids option causes the NFS server to ignore the auxiliary gids
from the incoming request. Instead the server looks up the the user's
supplementary groups from the local system. This is intended to support
users with more than 16 groups (as the NFS protocol only allows a maximum
of 16).

However, this does not make sense on the proxy, as the proxy is unaware of
users and groups. To support this feature the proxy would need to be
connected to something like LDAP.

The side-effect of this is that the incoming request has all its auxiliary
groups removed. This can cause permission errors when the proxy sends a
request to the source server.