Use fixed ports for NFS services
This allows defining a firewall rule that only permits clients to access
the ports required by NFS. Previously the firewall rule had to allow any
port.

This also improves stability when the load balancer connects a client to
a different instance. This was especially problematic if the client used
UDP to access the portmapper or mountd services, as the UDP load
balancer could connect the client to a different instance compared to
the TCP load balancer.

Change-Id: I8b46fdc8f454b045877f95ae8ecc743aa598d3c0