cachefilesd sometimes erroneously fails on start up
Sometimes when starting cachefilesd, systemd will report that cachefilesd
failed to start because it exited with an error.
However, /proc/fs/fscache/caches will show that the cache is active.

Running the start command a second time resolves the issue and shows in
systemd that the cache is running normally.

Change-Id: Id9900f8a217487f5168a45cc03d4cc9437385e31