Pause after rebooting image
Sometimes the script immediately after the reboot fails with the error
that the script could not be found. Suspect this is a race condition
between packer uploading the script and the machine rebooting.

Add a 30 second pause after the reboot step to try and reduce the
frequency of this error.

Change-Id: I4df016a41ee75240de1df36ac91d57ab26e3819b