Increase how much space is culled by cachefilesd
Increase the frun and brun limits from 10% to 20%. This causes
cachefilesd to reclaim more space once culling begins. The goal is to
reduce how often the watchdog needs to drop the slab cache when the
cache is full, as this can cause an increase in metadata operations that
need to be sent to the source server for data that is already in the
cache.

Change-Id: I0402588e277724e6f414b1b8000ef2cdc4c53bc0