Use latest 5.13 HWE kernel
The HWE APT packages only keep the binaries for the latest HWE build.
As such it's not possible to pin to a specific kernel version as every
6 weeks a new kernel will be released breaking the current script.

For now install the latest 5.13 kernel version. This will break at some
point, but by then Ubuntu 22.04 LTS should be tested and in use.

Change-Id: I3f17b9fbb0d9e06bcf4b28b4dc0de6196222b6f6