Update dashboard to use new metrics
This adds new graphs using the new metrics to show the total read/write
throughput between:

* KNFSD Proxy and Source.
* Clients and KNFSD Proxy.

The new dashboard also corrects an issue where the total number of
operations from the KNFSD Proxy to the Source were being under reported.
This is because the metric agent only parses a single `xprt` (transport)
entry.

Change-Id: I2a570b1d5cf1946808b1bd3ad8e5105476272e97