De-duplicate NFS stats from mounts sharing same NFS client
When multiple mounts share the same NFS client, only count their stats
once.

It was not possible to perform this aggregation in dashboards as the
dashboard has no way to filter which stats are duplicated due to sharing
and NFS client, and which stats are unique and need to be added.

Change-Id: Idf2deb4dc01dc04766d3ed416ee7d33871726da6