Use LF line endings for bash scripts and Linux config files
Create a .gitattributes file to set the line endings to LF for bash
scripts and Linux config files.

When checking out on other operating systems such as Windows the line
endings may be changed to CRLF. This will cause errors for bash scripts
and config files that require LF line endings.

Change-Id: Ib2e9e293dde332f13c95816a44200df4b6af165c