Support excluding servers when query_proxy_instance is enabled
Support a list of excluded servers and/or local paths when
query_proxy_instances is enabled.

When collecting metrics from clients with query_proxy_instances enabled
by default the collector will probe every NFS server that is mounted by
the client. This can cause issues if the client is mounting a mixture of
knfsd proxies and other NFS servers.

The metrics collector now has an exclude configuration section for
query_proxy_instances.

Change-Id: Iab25555f20b006b6dce02ef77344b613bcbc74c7