Add ability to route metrics over predictable IP ranges
This change adds a new ROUTE_METRICS_PRIVATE_GOOGLEAPIS config variable.

When set to `true`, the new `ROUTE_METRICS_PRIVATE_GOOGLEAPIS` variable will trigger an addition to the `/etc/hosts` file of each Knfsd Node for the following FQDN's:

* monitoring.googleapis.com
* logging.googleapis.com
* cloudtrace.googleapis.com

The IP used (`199.36.153.11`) is from the range defined in the Private Google Access docs. This ensures that metrics and logs are shipped over a predictable IP address range that is only routable from within Google Cloud.

Change-Id: I25b1742fa742074ab9144c94f582fd7cd2f574cc