Stop pinning to specific package versionsUpdate nfs-kernel-server version to 1:2.6.1-1ubuntu1.2
Pinning is causing issues with maintaining the repository and building
images because the apt repository only keeps the latest version for many
of these packages. This causes the image for release tags to stop
building when the package they depend upon is removed from the apt
repository.

Most of the packages are already unpinned and install the latest
version. The only components where the version really matters is the
kernel and nfs-tools, which are both installed separately from apt.

Change-Id: Ibb2d548322d08d1e95d1f59eefa62e933882be25