Change the default build machine type to c2-standard-16
This improves the build times for the custom kernel. These are generally
available and tend to offer the best build speed to price. Some
additional speed may be gained by using a c2-standard-30.

Can switch back to a smaller machine type once we're no longer building
a custom kernel.

Change-Id: Ie2369eb509ded27f4844f44964b7424d2aee4f81