Support deploying a cluster using stateful IP addresses
Stateful IPs are required to support load balancing using DNS round
robin. This is because clients will be connecting directly to a specific
proxy instances via the instance's internal IP address.

If a proxy instance fails and is replaced, the new instance needs to
keep the same IP address as the original so that clients can reconnect.

Change-Id: I3646cdee6ab534e00d919abc87e044c544f6c103