Temporary patch for cachefilesd issue
The cachefilesd service keeps terminating causing the cache to be
withdrawn. This results in the proxy no longer caching any data and just
acting as a pass-through server.

The underling issue is due to an old bug check that is no longer valid
when cachefilesd is compiled for 64-bit.

This patch should workaround the issue for the time being until the
proper fix has been decided by the cachefs maintainers.

https://listman.redhat.com/archives/linux-cachefs/2023-February/007449.html

Change-Id: I26bf6832bc503a3cf6266b0e0ac3032314a87d20