Add validation to terraform for FSID_MODE and AUTO_REEXPORT
Without the validation the Terraform can apply even if these values
are configured incorrectly as they're passed through as metadata to the
Compute Engine instance.

This results in either the proxy instance failing to start up and
getting into a reboot loop, or worse starting up but then failing to
function correctly.

Using Terraform 1.2 preconditions can be added to the resource's
lifecycle to verify the inputs before applying changes to the resource.

Change-Id: If7fb8df27076d7b0e5ebafb0da4d5801dcceabde