Document common errors deploying multiple clusters
When deploying multiple knfsd proxy clusters some of the resources
require globally unique names. Document the common Terraform error
messages when these names are not unique, and how to resolve the issue.

Added an error message to the startup script if the script terminates
abnormally. This makes it easier when looking at the logs to determine
if the script has terminated or is still running.

Change-Id: I98ed87f9ad3040931e6ea6e039d28e11b79bfba5