Fix scaling to zero when using round robin DNS
The A record always requires at least one entry. Remove the A record
from the DNS zone when scaling the proxy cluster to zero. Scaling can be
useful if you only want to temporarily shutdown (or restart) the cluster
without having to destroy all the other resources (such as Cloud SQL)
that can take several minutes to re-create.

Change-Id: I77b90384677c325253a1607139c6d3671bb5a681