Fix metrics agent config test not recognising external fsidd metrics
The config test for the metrics agent is aimed to detect if the config
file has not been updated when adding or removing a metric. Each metric
should have a corresponding transform in the config file.

Currently the test is failing because the config file contains
transforms for the knfsd-fsidd service. These metrics are not generated
by the metrics agent, as they're pushed to the metrics agent from the
knfsd-fsidd service.

Change-Id: I9572c3676b095013edc60a1108e157696c505059