Use terratest to execute the smoke tests
Terratest and Go provide a simple way to automate deploying the
Terraform config required by the test, then execute the tests.

The tests are now split into two parts, the terratest harness that is
executed on the Cloud Build docker image, and a remote component that
executes on the NFS client.

Change-Id: I4bb5282e63f7a3083ce1ea8164a583304ead6cae