Add endpoints to fetch NFS metrics from the knfsd-agent
This allows fetching instantaneous metrics from knfsd proxy instances
without needing to connect to the instance using SSH. These metrics are
not cached, and are queried directly from /proc/ on demand which makes
them useful for testing and troubleshooting.

Because some of the responses can get quite large (e.g. mountstats
returns about 300 lines of JSON per mount) the response will always be
gziped.

However, the data size is not considered a concern at the moment as
these endpoints are not intended for standard use. For standard
reporting of metrics the knfsd-metrics-agent should be configured to
send metrics periodically to an external system such as Cloud Monitoring
or Prometheus.

Change-Id: Icd57229e8de7f2aea8fe825d1091dfb72023eec1