Script to update all the GitHub refs in the documentation
It's getting hard to find all the places where the ref tags need
updating. This script will find any ref tag for
github.com/GoogleCloudPlatform/knfsd-cache-utils and update it.

The ./docs/changes/ directory is excluded from this replacement as the
history should be considered immutable. The changes refer to specific
releases at the time the change was made.

Also updated a few references in the documentation that were missed.

Change-Id: I6010c1d92000ca2f646ab0d1ee8d044ff9d1e612