Fix auto-reexporting the root of an NFS v4 server
When exporting the root "/" of an NFS v4 server with AUTO_REEXPORT
enabled; attempting to navigate to nested volumes would result in an
error such as "/mnt/files is not a directory".

This is because the logic that ensures the root export always has
"fsid=0" was overriding the re-export logic. The updated logic correctly
applies both conditions to the export.

Change-Id: Iaa47fc46df16281e91037405b5d1c332b779e995