Terraform module to configure Cloud Build
Provide a Terraform module that configures Cloud Build with a private
worker pool, and VPC network to support building and testing KNFSD proxy
images and Terraform configuration.

This module is opinionated to keep it simple, and can either be used as
is, or as an example of what resources need to be configured.

Change-Id: Ifd698cb87429e588303f1b0c12a448e728072a40