Removed SUBNETWORK_PROJECT (using self links instead)
Changed the format for shared VPC to use ID or self link instead.
This is because some resources such as Cloud SQL always require self
links, and other resources require self links when using shared VPC.

* Network ID format:
  projects/{{project}}/global/networks/{{name}}

* Subnetwork ID format:
  projects/{{project}}/regions/{{region}}/subnetworks/{{name}}

Simple names such as "default" are still supported. They will be
converted automatically to a network/subnetwork ID. The project will be
assumed to be the same project as the knfsd proxy cluster.

Change-Id: I8cb44038d69654a23f1e3bac6e981f5618abc724