Update to Ubuntu 24.04 LTS (Noble Numbat)
Update the image to use Ubuntu 24.04 LTS (Noble Numbat). Ubuntu 24.04
ships with the 6.8 kernel. The 6.8 kernel includes the kernel patches
that previously required compiling a custom kernel.

The 6.8 kernel has some bugs that affect CacheFiles when re-exporting.
This results in a large number of I/O errors, and can cause ext4 to
re-mount the cache file system as read-only.

The latest (6.11.0 at this time) HWE kernel for Ubuntu 24.04 has fixed
issues with CacheFiles that were present in the 6.8 kernel.

Change-Id: I97ec2b0a00537135dd9e18a6296ad5791203b64a