Disable unattended upgrades
The latest available HWE kernel (6.11.0) at the of this patch has an
issue where NFS can crash causing a kernel panic when the NFS server is
stopped due to a race condition. The main process that can restart the
NFS server is the unattended-upgrades.service. To avoid this issue this
service needs to be disabled.

This issue is fixed in later kernel versions (tested with 6.14.6
mainline). These fixes are not yet available in a supported HWE kernel
on Ubuntu 24.04.

Change-Id: I26d481fc991c490278680e0f2465d71e5496374d